requirements.txt
python-telegram-bot[job-queue]
requests
python-dotenv
flask
httpx